186: extract categories service
