+++ backend-rails/app/controllers/api/v1/categories_controller.rb
@@ -2,22 +2,22 @@ module Api
   module V1
     class CategoriesController < Api::V1::ApiController
       def index
-        categories = CategoriesFinderService.new(user: current_user, find_params:).call
+        categories = Categories::CategoriesFinderService.new(user: current_user, find_params:).call
         render json: CategoryPresenter.payload_for_list(categories)
       end
 
       def create
-        category = CategoryCreatorService.new(user: current_user, create_category_params:).call
+        category = Categories::CategoryCreatorService.new(user: current_user, create_category_params:).call
         render json: CategoryPresenter.payload_for_item(category), status: :created
       end
 
       def update
-        category = CategoryUpdaterService.new(user: current_user, update_category_params:).call
+        category = Categories::CategoryUpdaterService.new(user: current_user, update_category_params:).call
         render json: CategoryPresenter.payload_for_item(category), status: :ok
       end
 
       def destroy
-        CategoryDestroyerService.new(user: current_user, destroy_params:).call
+        Categories::CategoryDestroyerService.new(user: current_user, destroy_params:).call
         render json: {}, status: :no_content
       end
