+++ backend-rails/spec/services/categories/category_destroyer_service_spec.rb
@@ -2,26 +2,23 @@
 
 require 'rails_helper'
 
-RSpec.describe CategoryUpdaterService do
+RSpec.describe Categories::CategoryDestroyerService do
   describe '#call' do
     let!(:user) { create(:user) }
-    let!(:new_category_name) { create(:user) }
-    let(:update_category_params) { { id: category.id, name: new_category_name } }
+    let!(:destroy_params) { { id: category.id } }
 
-    subject { described_class.new(user:, update_category_params:).call }
+    subject { described_class.new(user:, destroy_params:).call }
 
     describe 'when user is category owner' do
       let!(:category) { create(:category, user:) }
-
-      it 'updates category name' do
-        expect { subject }.to change { category.reload.name }
+      it 'destroys category' do
+        expect { subject }.to change { Category.count }.from(1).to(0)
       end
     end
 
     describe 'when user is not category owner' do
       let!(:category) { create(:category) }
-
-      it 'raises not authorized error' do
+      it 'destroys category' do
         expect { subject }.to raise_error Pundit::NotAuthorizedError
       end
     end
