+++ backend-rails/spec/services/categories/category_updater_service_spec.rb
@@ -2,23 +2,26 @@
 
 require 'rails_helper'
 
-RSpec.describe CategoryDestroyerService do
+RSpec.describe Categories::CategoryUpdaterService do
   describe '#call' do
     let!(:user) { create(:user) }
-    let!(:destroy_params) { { id: category.id } }
+    let!(:new_category_name) { create(:user) }
+    let(:update_category_params) { { id: category.id, name: new_category_name } }
 
-    subject { described_class.new(user:, destroy_params:).call }
+    subject { described_class.new(user:, update_category_params:).call }
 
     describe 'when user is category owner' do
       let!(:category) { create(:category, user:) }
-      it 'destroys category' do
-        expect { subject }.to change { Category.count }.from(1).to(0)
+
+      it 'updates category name' do
+        expect { subject }.to change { category.reload.name }
       end
     end
 
     describe 'when user is not category owner' do
       let!(:category) { create(:category) }
-      it 'destroys category' do
+
+      it 'raises not authorized error' do
         expect { subject }.to raise_error Pundit::NotAuthorizedError
       end
     end
